keep-alive: drop unused nm_keep_alive_set_forced()
set-forced is currently unused, so drop it.

NMKeepAlive in principle determines the alive-status based on multiple
aspects, that in combination render the instance alive or dead. These
aspects cooperate in a particular way.

By default, a keep-alive instance should be alive. If there are conditions
enabled that further determine the alive-state, then these conditions
cooperate in a particular way. As it was, the force-flag would just
overrule them all.

But is that useful? The nm_keep_alive_set_forced() API also means that only
one user caller can have control over the flag. Independent callers cannot
cooperate on setting the flag, because there is no reference-counting or
registered handles.

At least today, it's unclear whether this flag really should overrule all
other conditions and how this flag would actually be used. Drop it for
now.