rh-utils: accept version number as argument for find-backports.sh
find-backports.sh only works because we craft commit messages with
necessary information. In particular the "Fixes" and cherry-picked-from
messages. That means, it relies on our git history to maintained in
a suitable manner so that the script can gather the necessary
information.

Likewise, we have a particular scheme how we do releases, how versions
are numbered, how stable branches and release tags are called, etc.

Exploit that, to allow for simpler calling convention for
find-backports.sh script:

  $ contrib/rh-utils/find-backports.sh 1.14

will automatically complete to

  $ contrib/rh-utils/find-backports.sh 1.14.0 nm-1-14 master