keep-alive: only declare keep-alive dead when connection becomes invisible
A user can always manually activate his/her own profile, even if the
profile is currently is invisible.

-- it could be invisible, because the profile has "connection.permissions"
set but the user has no active session.

Now, if the user should be able to activate such a profile, then it
cannot fail simply because being invisible all along. Instead, the
alive check must only fail, if a connection becomes invisible that was
visible in the time since it is monitored.